feat: Refactor type imports to use shared types and add GarageDetail type definition
