+++ apps/client/src/shared/types.ts
